refactor - index file added for each folder
