+++ src/components/Reusable/index.js
@@ -1,0 +1,21 @@
+import Text from './Text'
+import Button from './Button'
+import Card from './Card'
+import Header from './Header'
+import Icon from './Icon'
+import Image from './Image'
+import Input from './Input'
+import Main from './Main'
+import NavBar from './NavBar'
+
+export {
+    Text,
+    Button,
+    Card,
+    Header,
+    Icon,
+    Image,
+    Input,
+    Main,
+    NavBar,
+}
