+++ src/components/Wishlist/index.js
@@ -1,0 +1,11 @@
+import Wishlist from './Wishlist'
+import WishlistHeader from './WishlistHeader'
+import WishlistSection from './WishlistSection'
+import WishlistCard from './WishlistCard'
+
+export {
+    Wishlist,
+    WishlistSection,
+    WishlistHeader,
+    WishlistCard
+}
