+++ src/data/index.js
@@ -1,0 +1,5 @@
+import { wishlistImgArr } from './wishlist.img'
+
+export {
+    wishlistImgArr
+}
